added

+++ lib/screens/basic/screen_add_item.dart
@@ -545,7 +545,7 @@ class _ScreenAddItemState extends State<ScreenAddItem> {
         'cancel'.text().textButton(onPressed: () => Get.back()),
         'save'.text().textButton(
           onPressed: () {
-            final quantity = int.tryParse(itemCountController.text.trim());
+            final quantity = double.tryParse(itemCountController.text.trim());
             if (quantity == null || quantity <= 0) {
               Toaster.showError("Invalid quantity");
               return;
