+++ lib/screens/inventory/screen_add_transfer_order.dart
@@ -224,7 +224,7 @@ class _ScreenAddTransferOrderState extends State<ScreenAddTransferOrder> {
         'update'.text().textButton(
           onPressed: () {
             try {
-              int quantity = int.parse(proposedQuantity.text);
+              double quantity = double.parse(proposedQuantity.text);
               double cost = double.parse(proposedCostsPrice.text);
               itemInv.quantity = quantity;
               itemInv.cost = cost;
